feat: Add GCP Cloud SQL for PostgreSQL support (#51)
* FEATURE: Add GCP Cloud SQL for PostgreSQL support

* Add Guardrails and Config updates for Cloud SQL

* Allow config to be pulled from env

Allows the GCP config to pulled from the environment. Also add
validation to the config to ensure that the GCP fields are present for a
cloudsql agent. Add in ability to detect a cloud sql config without the
explicit flag.

Separately, filter out config knobs marked as internal by cloud sql.
These seem to be able to change at random and are outside of our control
so we ignore them.

* Wait for updates to be applied

We now wait for the update operation created in GCP to be DONE before acknowledging our update as successful. From testing so far, this is around 10 seconds or so of downtime.

* Improve handling of possible error scenarios in `GetMemoryMetrics`

Make sure that the filters we provide only ever get one result so that we
are more confident that the agent is looking at the right db,

Also make sure that we never try and read data that we don't have from the response
and so avoid a runtime panic.