Use pg max connections
The value for max connections gets set by azure based on the SKU (previous calcualted value in the agent) as a default, but it can be changed after the server creation. This fact, combined with the fact that it does not get changed by azure if you upgrade your server means that the value we are using currently may be wrong (leading to incorrect work_mem search space defintions) this commit changes that.